+++ actix-multipart/CHANGES.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+- Include inner error in `MultipartError::Payload`, `MultipartError::Parse`, and `MultipartError::Field` display strings. [#4168]
+- Fix `MultipartError::ContentDispositionNameMissing` display string to indicate missing name parameter.
+- Improve `JsonFieldError::ContentType` and `TextError::ContentType` error messages.
+
+[#4168]: https://github.com/actix/actix-web/issues/4168
+
 ## 0.8.0
 
 - Add multi-field multipart payload builders to `actix_multipart::test`. [#3575]