+++ actix-multipart/CHANGES.md
@@ -2,10 +2,12 @@
 
 ## Unreleased
 
+- Fix multipart field parsing when boundary delimiter is split across payload chunks. [#3635]
 - Include inner error in `MultipartError::Payload`, `MultipartError::Parse`, and `MultipartError::Field` display strings. [#4168]
 - Fix `MultipartError::ContentDispositionNameMissing` display string to indicate missing name parameter.
 - Improve `JsonFieldError::ContentType` and `TextError::ContentType` error messages.
 
+[#3635]: https://github.com/actix/actix-web/issues/3635
 [#4168]: https://github.com/actix/actix-web/issues/4168
 
 ## 0.8.0