+++ actix-multipart/CHANGES.md
@@ -2,10 +2,10 @@
 
 ## Unreleased
 
-- Fix multipart field parsing when boundary delimiter is split across payload chunks. [#3635]
-- Improve detail in some error messages.
+## 0.8.1
 
-[#3635]: https://github.com/actix/actix-web/issues/3635
+- Fix multipart field parsing when boundary delimiter is split across payload chunks.
+- Improve detail in some error messages.
 
 ## 0.8.0